# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis Inspector",
    "description": "\nThe inspection and governance engine behind the **Aegis** Zero-Trust\nResponsible AI Gateway.\n\nThis service decides whether a prompt may be transmitted, screens model\nresponses on the way back, and records the platform's decision history. It is\nconsumed by the Aegis Gateway, never by end clients directly.\n\n### What this service does not claim\n\n* Detection is **heuristic and incomplete**. Prompt-injection defence covers\n  known OWASP LLM01 patterns and is bypassable by obfuscation.\n* Grounded Response Verification is a **support score**, not a guarantee of\n  factual correctness.\n* Energy and CO2 figures are **estimates** from configurable assumptions, not\n  measurements.\n* Audit output is **transaction evidence** supporting a deployer's own\n  record-keeping. It is not a conformity assessment and does not establish\n  legal compliance.\n",
    "version": "0.1.0"
  },
  "paths": {
    "/inspect": {
      "post": {
        "tags": [
          "inspection"
        ],
        "summary": "Inspect an outbound prompt and return a policy decision",
        "description": "Stateless. Returns ALLOW / SANITIZE / WARN / BLOCK plus the sanitized messages, the placeholder-to-original vault map, cache and routing hints, and measured per-stage timings.\n\nAlways returns HTTP 200 -- a BLOCK is a *decision*, not a transport error. The Gateway is responsible for translating `block_reason.http_status` into the status code the client sees.",
        "operationId": "inspect_inspect_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InspectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InspectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/inspect/egress": {
      "post": {
        "tags": [
          "inspection"
        ],
        "summary": "Screen a model response before it reaches the user",
        "description": "Runs harm and bias screening, and grounded verification when `reference_context` is supplied. Returns PASS / ANNOTATE / REPLACE.",
        "operationId": "inspect_egress_inspect_egress_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EgressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EgressResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/verify": {
      "post": {
        "tags": [
          "inspection"
        ],
        "summary": "Grounded Response Verification (standalone)",
        "description": "Compares an answer against a reference document using sentence-level claim extraction and NLI entailment. This is a heuristic support score, not a guarantee of factual correctness.",
        "operationId": "verify_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/policies": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Current policy config",
        "operationId": "get_policies_api_policies_get",
        "parameters": [
          {
            "name": "profile",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "governance"
        ],
        "summary": "Update policy, creating a new immutable version",
        "description": "Policy changes are themselves auditable. Each update writes a new `policy_version` row; previous versions are never mutated, so any past decision can be replayed against the rules that were in force.",
        "operationId": "put_policies_api_policies_put",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reviews": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Appeal an automated block",
        "description": "Aegis makes automated decisions that affect people, which makes Aegis itself subject to the oversight principle it enforces. Any appealable block can be contested here with a written justification.",
        "operationId": "create_review_api_reviews_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Review queue",
        "operationId": "list_reviews_api_reviews_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReviewStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reviews/{review_id}/decision": {
      "post": {
        "tags": [
          "governance"
        ],
        "summary": "Human approves or denies an appeal",
        "description": "Approval mints a short-lived, single-use override token scoped to the one `request_id` under review. The Gateway replays that request with the token; `/inspect` honours it once and records the override.",
        "operationId": "decide_review_api_reviews__review_id__decision_post",
        "parameters": [
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fairness/report": {
      "get": {
        "tags": [
          "governance"
        ],
        "summary": "Detector recall across population groups",
        "description": "Reports measured PII-detector recall per name-origin group, and the gap between the best- and worst-served group.\n\nA privacy tool that protects some people's identities more reliably than others is an unfair system. Aegis measures its own gap and publishes it, including the baseline where it was worse.\n\nReturns nulls until Phase 13 -- the numbers are measured, never assumed.",
        "operationId": "fairness_report_api_fairness_report_get",
        "parameters": [
          {
            "name": "detector",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pii.person",
              "title": "Detector"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FairnessReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EgressRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "default"
          },
          "response_text": {
            "type": "string",
            "title": "Response Text"
          },
          "reference_context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Context"
          },
          "checks": {
            "items": {
              "type": "string",
              "enum": [
                "harm",
                "bias",
                "grounding"
              ]
            },
            "type": "array",
            "title": "Checks"
          },
          "policy_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Profile"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "response_text"
        ],
        "title": "EgressRequest"
      },
      "EgressResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "safety": {
            "$ref": "#/components/schemas/SafetyResult"
          },
          "bias": {
            "$ref": "#/components/schemas/BiasResult"
          },
          "grounding": {
            "$ref": "#/components/schemas/GroundingResult"
          },
          "action": {
            "$ref": "#/components/schemas/EgressAction",
            "default": "PASS"
          },
          "replacement_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replacement Text"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            },
            "type": "array",
            "title": "Pipeline"
          },
          "total_duration_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Duration Ms",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "EgressResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InspectRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "default"
          },
          "user_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Ref"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages"
          },
          "mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "policy_profile": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Profile"
          },
          "locale_hints": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Locale Hints"
          },
          "override_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Token"
          },
          "options": {
            "$ref": "#/components/schemas/InspectOptions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "messages"
        ],
        "title": "InspectRequest"
      },
      "InspectResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "policy_version": {
            "type": "integer",
            "title": "Policy Version",
            "default": 1
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "block_reason": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BlockReason"
              },
              {
                "type": "null"
              }
            ]
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "Messages"
          },
          "detections": {
            "$ref": "#/components/schemas/DetectionBundle"
          },
          "counts": {
            "$ref": "#/components/schemas/DetectionCounts"
          },
          "vault": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Vault"
          },
          "vault_policy": {
            "$ref": "#/components/schemas/VaultPolicy"
          },
          "cache": {
            "$ref": "#/components/schemas/CacheHint"
          },
          "routing_hint": {
            "$ref": "#/components/schemas/RoutingHint"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation",
            "default": ""
          },
          "override_applied": {
            "type": "boolean",
            "title": "Override Applied",
            "default": false
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            },
            "type": "array",
            "title": "Pipeline"
          },
          "total_duration_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Duration Ms",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "decision",
          "messages"
        ],
        "title": "InspectResponse"
      },
...
      "PolicyResponse": {
        "properties": {
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "available_profiles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Available Profiles"
          },
          "yaml_body": {
            "type": "string",
            "title": "Yaml Body"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "version",
          "profile",
          "yaml_body"
        ],
        "title": "PolicyResponse"
      },
      "PolicyUpdateRequest": {
        "properties": {
          "yaml_body": {
            "type": "string",
            "title": "Yaml Body"
          },
          "author_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Ref"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "yaml_body"
        ],
        "title": "PolicyUpdateRequest"
      },
...
      "ReviewCreateRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "default"
          },
          "user_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Ref"
          },
          "original_decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "rule_fired": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Fired"
          },
          "user_justification": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "User Justification"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "original_decision",
          "user_justification"
        ],
        "title": "ReviewCreateRequest"
      },
      "ReviewCreateResponse": {
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewRecord"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review",
          "accepted"
        ],
        "title": "ReviewCreateResponse"
      },
...
      "VerifyRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "reference_context": {
            "type": "string",
            "title": "Reference Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "answer",
          "reference_context"
        ],
        "title": "VerifyRequest",
        "description": "Standalone grounding check -- POST /verify."
      },
      "VerifyResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "grounding": {
            "$ref": "#/components/schemas/GroundingResult"
          },
          "pipeline": {
            "items": {
              "$ref": "#/components/schemas/PipelineStage"
            },
            "type": "array",
            "title": "Pipeline"
          },
          "total_duration_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Total Duration Ms",
            "default": 0.0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "grounding"
        ],
        "title": "VerifyResponse"
      }
    }
  },
...
}